refactor(userspace/libscap): isolate scap reader definitions
