+++ userspace/libscap/scap-int.h
@@ -32,6 +32,7 @@ limitations under the License.
 #include "plugin_info.h"
 #include "scap_assert.h"
 #include "scap_zlib.h"
+#include "scap_reader.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -64,33 +65,6 @@ typedef struct scap_tid
 	UT_hash_handle hh; ///< makes this structure hashable
 } scap_tid;
 
-typedef enum ppm_reader_type
-{
-	RT_FILE = 0,
-	RT_BUFFERED = 1,
-} ppm_reader_type;
-
-struct scap_reader
-{
-	ppm_reader_type m_type;
-	union
-	{
-		struct // RT_FILE
-		{
-			gzFile m_file; ///< The file to read data from
-		};
-		struct // RT_BUFFERED
-		{
-			bool m_free_reader; ///< whether the reader should be free-d on close
-			bool m_has_err; ///< True if the most recent m_reader operation had an error
-			uint8_t* m_buffer; ///< The buffer used to read data from m_reader
-			uint32_t m_buffer_cap; ///< The physical size of the buffer
-			uint32_t m_buffer_len; ///< The number of bytes used in the buffer
-			uint32_t m_buffer_off; ///< The cursor position in the buffer
-			struct scap_reader* m_reader; ///< The reader to read from in buffered mode
-		};
-	};
-};
 
 //
 // The open instance handle
@@ -352,185 +326,6 @@ void udig_end_capture(struct scap_engine_handle engine);
 int32_t udig_set_snaplen(struct scap_engine_handle engine, uint32_t snaplen);
 int32_t udig_stop_dropping_mode(struct scap_engine_handle engine);
 int32_t udig_start_dropping_mode(struct scap_engine_handle engine, uint32_t sampling_ratio);
-
-//
-// scap_reader functions implementation
-//
-
-static inline scap_reader_t *scap_reader_open_gzfile(gzFile file)
-{
-	if (file == NULL)
-	{
-		return NULL;
-	}
-	scap_reader_t* r = (scap_reader_t *) malloc (sizeof (scap_reader_t));
-	r->m_type = RT_FILE;
-	r->m_file = file;
-	return r;
-}
-
-// if own_reader is true, the wrapped reader will be de-allocated using free()
-// when the buffered reader gets closed 
-static inline scap_reader_t *scap_reader_open_buffered(scap_reader_t* reader, uint32_t bufsize, bool own_reader)
-{
-	if (reader == NULL || bufsize == 0)
-	{
-		return NULL;
-	}
-	scap_reader_t* r = (scap_reader_t*) malloc (sizeof(scap_reader_t));
-	r->m_type = RT_BUFFERED;
-	r->m_free_reader = own_reader;
-	r->m_has_err = false;
-	r->m_reader = reader;
-	r->m_buffer = (uint8_t*) malloc (sizeof(uint8_t) * bufsize);
-	r->m_buffer_cap = bufsize;
-	r->m_buffer_len = 0;
-	r->m_buffer_off = 0;
-	return r;
-}
-
-static inline ppm_reader_type scap_reader_type(scap_reader_t *r)
-{
-	ASSERT(r != NULL);
-	return r->m_type;
-}
-
-static inline int scap_reader_read(scap_reader_t *r, void* buf, uint32_t len)
-{
-	ASSERT(r != NULL);
-	uint8_t* buf_bytes = (uint8_t*) buf;
-	switch (r->m_type)
-	{
-		case RT_FILE:
-			return gzread(r->m_file, buf, len);
-		case RT_BUFFERED:
-			while (len > 0 && !r->m_has_err)
-			{
-				if (r->m_buffer_off >= r->m_buffer_len)
-				{
-					int nread = scap_reader_read(r->m_reader, r->m_buffer, r->m_buffer_cap);
-					if (nread <= 0)
-					{
-						// invalidate next read
-						r->m_has_err = true;
-						return buf_bytes - (uint8_t*) buf;
-					}
-					r->m_buffer_off = 0;
-					r->m_buffer_len = (uint32_t) nread;
-				}
-				*(buf_bytes++) = r->m_buffer[r->m_buffer_off++];
-				len--;
-			}
-			return buf_bytes - (uint8_t*) buf;
-		default:
-			ASSERT(false);
-			return 0;
-	}
-}
-
-static inline int64_t scap_reader_offset(scap_reader_t *r)
-{
-	ASSERT(r != NULL);
-	switch (r->m_type)
-	{
-		case RT_FILE:
-			return gzoffset(r->m_file);
-		case RT_BUFFERED:
-			return scap_reader_offset(r->m_reader);
-		default:
-			ASSERT(false);
-			return -1;
-	}
-}
-
-static inline int64_t scap_reader_tell(scap_reader_t *r)
-{
-	int64_t res;
-	ASSERT(r != NULL);
-	switch (r->m_type)
-	{
-		case RT_FILE:
-			return gztell(r->m_file);
-		case RT_BUFFERED:
-			res = scap_reader_tell(r->m_reader);
-			if (res < 0)
-			{
-				return res;
-			}
-			return res - r->m_buffer_len + r->m_buffer_off;
-		default:
-			ASSERT(false);
-			return -1;
-	}
-}
-
-static inline int64_t scap_reader_seek(scap_reader_t *r, int64_t offset, int whence)
-{
-	ASSERT(r != NULL);
-	switch (r->m_type)
-	{
-		case RT_FILE:
-			return gzseek(r->m_file, offset, whence);
-		case RT_BUFFERED:
-			if (whence == SEEK_CUR)
-			{
-				if (offset < 0 && r->m_buffer_off >= (uint32_t) (offset * -1))
-				{
-					r->m_buffer_off -= (uint32_t) (offset * -1);
-					return scap_reader_tell(r->m_reader);
-				}
-				else if (offset > 0 && r->m_buffer_len >= r->m_buffer_off + (uint32_t) offset)
-				{
-					r->m_buffer_off += (uint32_t) offset;
-					return scap_reader_tell(r->m_reader);
-				}
-			}
-			r->m_buffer_off = 0;
-			r->m_buffer_len = 0;
-			return scap_reader_seek(r->m_reader, offset, whence);
-		default:
-			ASSERT(false);
-			return -1;
-	}
-}
-
-static inline const char *scap_reader_error(scap_reader_t *r, int *errnum)
-{
-	ASSERT(r != NULL);
-	switch (r->m_type)
-	{
-		case RT_FILE:
-			return gzerror(r->m_file, errnum);
-		case RT_BUFFERED:
-			return scap_reader_error(r->m_reader, errnum);
-		default:
-			ASSERT(false);
-			*errnum = -1;
-			return "unknown scap_reader type";
-	}
-}
-
-static inline int scap_reader_close(scap_reader_t *r)
-{
-	int res;
-	ASSERT(r != NULL);
-	switch (r->m_type)
-	{
-		case RT_FILE:
-			return gzclose(r->m_file);
-		case RT_BUFFERED:
-			res = scap_reader_close(r->m_reader);
-			if (r->m_free_reader)
-			{
-				free(r->m_reader);
-			}
-			free(r->m_buffer);
-			return res;
-		default:
-			ASSERT(false);
-			return -1;
-	}
-}
 
 #ifdef __cplusplus
 }
